requirements.txt
absl-py==0.13.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.9.3
blis==0.4.1
bs4==0.0.1
bz2file==0.98
cachetools==4.2.2
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
chex==0.0.8
chromedriver-install==1.0.3
click==8.0.3
cloudpickle==2.0.0
coloredlogs==14.0
cryptography==36.0.1
cycler==0.10.0
cymem==2.0.5
Cython==0.29.25
decorator==5.1.0
dill==0.3.4
dm-tree==0.1.6
dopamine-rl==3.2.1
# en-core-web-sm==3.1.0
et-xmlfile==1.1.0
faiss-cpu==1.7.1.post3
filelock==3.3.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
flatbuffers==1.12
flax==0.3.4
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.3.3
gevent==21.8.0
gin-config==0.4.0
google==3.0.0
google-api-core==2.0.1
google-api-python-client==2.21.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.32.0
gunicorn==20.1.0
gym==0.20.0
h11==0.13.0
h5py==2.10.0
httplib2==0.19.1
humanfriendly==10.0
idna==2.10
importlib-resources==5.2.2
install==1.3.5
itsdangerous==2.0.1
jax==0.2.20
jaxlib==0.1.71
Jinja2==3.0.2
joblib==1.1.0
Keras-Preprocessing==1.1.2
kfac==0.2.3
kiwisolver==1.3.2
lxml==4.6.5
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mesh-tensorflow==0.1.19
mpmath==1.2.1
msgpack==1.0.2
murmurhash==1.0.5
mysql-connector-python==8.0.26
nltk==3.6.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
openai==0.11.3
opencv-python==4.5.3.56
openpyxl==3.0.9
opt-einsum==3.3.0
optax==0.0.9
outcome==1.1.0
packaging==21.0
pandas==1.3.4
pandas-stubs==1.2.0.39
pathlib==1.0.1
pathy==0.6.0
Pillow==8.3.2
plac==1.1.3
preshed==3.0.5
promise==2.3
protobuf==3.18.1
# PyAFF==0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.5
pygaggle==0.0.3.1
pygame==2.0.1
pyjnius==1.4.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pypng==0.0.21
pyserini==0.10.1.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-Levenshtein==0.12.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.9.24
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
selenium==4.1.0
sentencepiece==0.1.94
simplejson==3.17.5
six==1.15.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==2.2.4
spacy-legacy==3.0.8
srsly==1.0.5
sympy==1.8
tensor2tensor==1.15.7
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-addons==0.14.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.4.0
tensorflow-gan==2.1.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.2.0
tensorflow-probability==0.7.0
termcolor==1.1.0
tf-slim==1.1.0
thinc==7.4.0
threadpoolctl==3.0.0
tokenizers==0.9.4
toolz==0.11.1
torch==1.4.0
tqdm==4.45.0
transformers==4.0.0
trio==0.19.0
trio-websocket==0.9.2
typeguard==2.12.1
typer==0.4.0
typing-extensions==3.7.4.3
Unidecode==1.3.4
uritemplate==3.0.1
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.1
Whoosh==2.7.4
wikipedia==1.4.0
wrapt==1.12.1
wsproto==1.0.0
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
